requirements.txt
django
gunicorn
pyjwt